bättre implementation av movement

+++ src/movingObject.java
@@ -1,6 +1,9 @@
 public class movingObject {
-    int x;
-    int y;
+    //x and y position of the object
+    float x;
+    float y;
+
+    //size of the object
     int width;
     int height;
     public movingObject(int x, int y, int width, int height){
